feat(browser): extract Design inspector View and slim layout palette
Split the adjustment panel into a fixture-driven View + store Container,
add Storybook stories/fixtures, slim the overlay layout palette to three
sections, default marker click to edit, and add a static overlay demo for
visual review.
