+++ apps/desktop/src/preload/browser-annotation-runtime.ts
@@ -154,9 +154,6 @@ const BROWSER_ANNOTATION_LAYOUT_COMPONENTS: ReadonlyArray<{
       { type: 'footer', label: 'Footer', ...DEFAULT_SIZES.footer },
       { type: 'modal', label: 'Modal', ...DEFAULT_SIZES.modal },
       { type: 'banner', label: 'Banner', ...DEFAULT_SIZES.banner },
-      { type: 'drawer', label: 'Drawer', ...DEFAULT_SIZES.drawer },
-      { type: 'popover', label: 'Popover', ...DEFAULT_SIZES.popover },
-      { type: 'divider', label: 'Divider', ...DEFAULT_SIZES.divider },
     ],
   },
   {
@@ -165,20 +162,9 @@ const BROWSER_ANNOTATION_LAYOUT_COMPONENTS: ReadonlyArray<{
       { type: 'card', label: 'Card', ...DEFAULT_SIZES.card },
       { type: 'text', label: 'Text', ...DEFAULT_SIZES.text },
       { type: 'image', label: 'Image', ...DEFAULT_SIZES.image },
-      { type: 'video', label: 'Video', ...DEFAULT_SIZES.video },
       { type: 'table', label: 'Table', ...DEFAULT_SIZES.table },
-      { type: 'grid', label: 'Grid', ...DEFAULT_SIZES.grid },
       { type: 'list', label: 'List', ...DEFAULT_SIZES.list },
-      { type: 'chart', label: 'Chart', ...DEFAULT_SIZES.chart },
-      { type: 'codeBlock', label: 'Code Block', ...DEFAULT_SIZES.codeBlock },
-      { type: 'map', label: 'Map', ...DEFAULT_SIZES.map },
-      { type: 'timeline', label: 'Timeline', ...DEFAULT_SIZES.timeline },
-      { type: 'calendar', label: 'Calendar', ...DEFAULT_SIZES.calendar },
-      { type: 'accordion', label: 'Accordion', ...DEFAULT_SIZES.accordion },
-      { type: 'carousel', label: 'Carousel', ...DEFAULT_SIZES.carousel },
-      { type: 'logo', label: 'Logo', ...DEFAULT_SIZES.logo },
-      { type: 'faq', label: 'FAQ', ...DEFAULT_SIZES.faq },
-      { type: 'gallery', label: 'Gallery', ...DEFAULT_SIZES.gallery },
+      { type: 'grid', label: 'Grid', ...DEFAULT_SIZES.grid },
     ],
   },
   {
@@ -191,47 +177,6 @@ const BROWSER_ANNOTATION_LAYOUT_COMPONENTS: ReadonlyArray<{
       { type: 'tabs', label: 'Tabs', ...DEFAULT_SIZES.tabs },
       { type: 'dropdown', label: 'Dropdown', ...DEFAULT_SIZES.dropdown },
       { type: 'toggle', label: 'Toggle', ...DEFAULT_SIZES.toggle },
-      { type: 'stepper', label: 'Stepper', ...DEFAULT_SIZES.stepper },
-      { type: 'rating', label: 'Rating', ...DEFAULT_SIZES.rating },
-      { type: 'fileUpload', label: 'File Upload', ...DEFAULT_SIZES.fileUpload },
-      { type: 'checkbox', label: 'Checkbox', ...DEFAULT_SIZES.checkbox },
-      { type: 'radio', label: 'Radio', ...DEFAULT_SIZES.radio },
-      { type: 'slider', label: 'Slider', ...DEFAULT_SIZES.slider },
-      { type: 'datePicker', label: 'Date Picker', ...DEFAULT_SIZES.datePicker },
-    ],
-  },
-  {
-    section: 'Elements',
-    items: [
-      { type: 'avatar', label: 'Avatar', ...DEFAULT_SIZES.avatar },
-      { type: 'badge', label: 'Badge', ...DEFAULT_SIZES.badge },
-      { type: 'tag', label: 'Tag', ...DEFAULT_SIZES.tag },
-      { type: 'breadcrumb', label: 'Breadcrumb', ...DEFAULT_SIZES.breadcrumb },
-      { type: 'pagination', label: 'Pagination', ...DEFAULT_SIZES.pagination },
-      { type: 'progress', label: 'Progress', ...DEFAULT_SIZES.progress },
-      { type: 'alert', label: 'Alert', ...DEFAULT_SIZES.alert },
-      { type: 'toast', label: 'Toast', ...DEFAULT_SIZES.toast },
-      { type: 'notification', label: 'Notification', ...DEFAULT_SIZES.notification },
-      { type: 'tooltip', label: 'Tooltip', ...DEFAULT_SIZES.tooltip },
-      { type: 'stat', label: 'Stat', ...DEFAULT_SIZES.stat },
-      { type: 'skeleton', label: 'Skeleton', ...DEFAULT_SIZES.skeleton },
-      { type: 'chip', label: 'Chip', ...DEFAULT_SIZES.chip },
-      { type: 'icon', label: 'Icon', ...DEFAULT_SIZES.icon },
-      { type: 'spinner', label: 'Spinner', ...DEFAULT_SIZES.spinner },
-    ],
-  },
-  {
-    section: 'Blocks',
-    items: [
-      { type: 'pricing', label: 'Pricing', ...DEFAULT_SIZES.pricing },
-      { type: 'testimonial', label: 'Testimonial', ...DEFAULT_SIZES.testimonial },
-      { type: 'cta', label: 'CTA', ...DEFAULT_SIZES.cta },
-      { type: 'productCard', label: 'Product Card', ...DEFAULT_SIZES.productCard },
-      { type: 'profile', label: 'Profile', ...DEFAULT_SIZES.profile },
-      { type: 'feature', label: 'Feature', ...DEFAULT_SIZES.feature },
-      { type: 'team', label: 'Team', ...DEFAULT_SIZES.team },
-      { type: 'login', label: 'Login', ...DEFAULT_SIZES.login },
-      { type: 'contact', label: 'Contact', ...DEFAULT_SIZES.contact },
     ],
   },
 ]
@@ -239,7 +184,7 @@ const BROWSER_ANNOTATION_LAYOUT_COMPONENTS: ReadonlyArray<{
 const BROWSER_ANNOTATION_DEFAULT_SETTINGS: BrowserAnnotationRuntimeSettings = {
   blockInteractions: true,
   clearOnCopySend: false,
-  markerClickBehavior: 'delete',
+  markerClickBehavior: 'edit',
   markerColorId: 'blue',
   outputDetail: 'detailed',
   reactDetectionEnabled: true,
