+++ apps/desktop/src/preload/demos/README.md
@@ -1,0 +1,11 @@
+# Preload demos
+
+Static visual fixtures under this directory are **not runtime code**. They exist for screenshots, design review, and Storybook-like visual inspection of Cradle UI chrome.
+
+## Files
+
+| File | Purpose |
+|------|---------|
+| `browser-annotation-overlay-demo.html` | Self-contained 1280×800 composition approximating the in-browser annotation overlay (toolbar, selection ring, marker badge, comment editor) over a mock product page. Open directly in a browser or embed in docs. |
+
+These pages use inline CSS only and have no build step or external dependencies.
