+++ apps/web/src/features/browser/README.md
@@ -19,8 +19,10 @@ The renderer publishes the active surface kind and bounds. Electron main adopts 
 - **browser-panel.logic.ts**: Pure address-bar logic copied from Synara; owns display value normalization, browser-style search/URL coercion, tab/history suggestions, chrome status, and address sync decisions.
 - **browser-panel-loader.ts**: Shared lazy loader and preload hook for the Electron browser panel.
 - **browser-panel.tsx**: Mixed right-side panel UI scoped to the owning route surface id; renders native browser tabs, a chooser launcher tab, TUI shell tabs, workspace file preview/editor tabs, workspace diff tabs, browser navigation controls, address suggestions, composer screenshot attachment, browser comment controls/tray, and a native viewport anchor whose bounds are synchronized to desktop `WebContentsView`. Active browser comment interaction is owned by the page preload runtime inside the real browser `WebContents`; renderer state stores runtime events, selected elements, draft design payloads, and screenshots captured only when saving or sending evidence.
-- **browser-annotation-adjustment-panel.tsx**: Right-aside Design/CSS inspector for the selected page element. Draft style edits live-preview through `applyAnnotationDesign` and commit as annotation design changes.
+- **browser-annotation-adjustment-panel-view.tsx**: Fixture-driven Design/CSS inspector View (Position / Layout / Dimensions / Spacing / Typography / Fill / Stroke / Effects). Storybook: `App/Browser/Annotation Adjustment Panel`.
+- **browser-annotation-adjustment-panel.tsx**: Container that wires `annotationAdjustmentSession` from the BrowserPanel store into the View and dispatches `browser:annotation-adjustment-apply`.
 - **browser-annotation-design-fields.ts**: Field model for the Design inspector (groups, segment options, scrub helpers). Mirrors the desktop preload design CSS property list.
+- **fixtures/browser-annotation-adjustment.ts**: Selected-element + draft-change fixtures for Storybook.
 - **browser-panel.test.tsx**: Regression tests for BrowserPanel shell render boundaries, workspace diff scroll commands, cross-session browser tab source markers, native browser tab request handling, and native surface suppression.
 - **index.ts**: Browser feature barrel export.
 - **native-surface-occlusion.ts**: Attribute helpers for DOM chrome that should trim the native browser viewport bounds.
